requirements.txt
streamlit
PyMuPDF
openai
python-dotenv